move testutil to analysis
